requirements.txt
Flask
torch
facenet-pytorch
opencv-python-headless
numpy
Pillow
